SDK folder added
