+++ ExampleSwift/ExampleSwift/Constants.swift
@@ -15,4 +15,5 @@ let PTEventParkingSuspected: String = "Parking Suspected"
 let PTEventVehicleParked: String = "Vehicle Parked"
 let PTEventSearchingParking: String = "Searching Parking"
 let PTEventUpdateLocation: String = "Update Location"
+let PTEventSTMP: String = "STMP Callback"
 let PTCatagoryName: String = "SDK Event"
