+++ ExampleSwift/Podfile
@@ -7,5 +7,5 @@ target 'ExampleSwift' do
 
   # Pods for TableViewCheck
 	#pod 'PredictIO-iOS', '~> 0.5.0'
-	pod 'PredictIO', '~> 0.1.0'
+	pod 'PredictIO', '~> 3.0.1'
 end
