+++ PredictIO.podspec
@@ -1,6 +1,6 @@
 Pod::Spec.new do |s|
   s.name             = 'PredictIO'
-  s.version          = '0.1.0'
+  s.version          = '3.0.0'
   s.license          = { :type => 'Apache License, Version 2.0', :file => 'LICENSE' }
   s.homepage         = 'https://github.com/predict-io/PredictIO-iOS'
   s.summary          = 'The parking detection API (patent pending) allows you to retrieve real-time updates on the parking status of a user.'
@@ -12,11 +12,12 @@ Pod::Spec.new do |s|
   s.ios.deployment_target = '7.0'
   s.requires_arc = true
 
-  s.source_files = 'PredictIO-iOS/Classes/**/*'
-  s.vendored_library = 'PredictIO-iOS/libPredictIO.a'
+  s.source_files = 'SDK/*'
+  s.public_header_files = 'SDK/*.h'
+  s.vendored_library = 'SDK/libParktagSDK.a'
 
   s.frameworks = 'UIKit', 'CoreMotion', 'CoreLocation', 'CoreTelephony', 'AdSupport', 'AVFoundation', 'CoreBluetooth', 'SystemConfiguration', 'ExternalAccessory'
 #2
-s.description      = "You may use the parking updates to trigger events or notifications specific to your use case. The parking detection service stack leverages all available phone sensors. Sensor usage is extremely battery optimized. Incremental battery consumption does not exceed 10% in typical cases. Also we strongly suggest you ensure inclusion of LBS in your T&C and Privacy Policies. By using this API you explicitly agree to our license agreement, terms and conditions and privacy policy. If you are unsure about any item, please contact us at support@parktag.mobi."
+  s.description      = "You may use the parking updates to trigger events or notifications specific to your use case. The parking detection service stack leverages all available phone sensors. Sensor usage is extremely battery optimized. Incremental battery consumption does not exceed 10% in typical cases. Also we strongly suggest you ensure inclusion of LBS in your T&C and Privacy Policies. By using this API you explicitly agree to our license agreement, terms and conditions and privacy policy. If you are unsure about any item, please contact us at support@parktag.mobi."
 
 end
